Adds more low-level optimizations for sparse probabilitiy calcs.
This commit focuses on the innermost loop, and makes the following
updates:
- new faster kron-product algorithm (in cython still)
- tweaks from profiling cython for embedding acton(...) logic
- new "special case2" embedded gate "acton" method (in cython).
- fixes can't-pickle-instance-method bug w/EmbeddedGateMap.acton
- try/except ignores lack of frobeniusdist implementations in Gate
   classes when comparing the GateSets on different processor ranks.

At this point, profiling 3,4,5,& 6-qubit data-generation examples
shows the desired result that everything that's taking up time
is necessary math in C/Fortran.

MPI distribution within GateMapCalc also seems to be working, as
1- vs 2-processor comparisons of a 3-qubit GST testcase were in
perfect agreement.

This commit leaves many debug statements and overall messiness,
particularly in gate.py and fastcalc.pyx, which should be cleaned
up in the future (look for "DEBUG", "REMOVE" and "OLD", and
consider better ordering of methods, splitting fastcalc, etc.).