Updates LindbladParameterizedGate to do basis transform earlier.
Previously, the *dense-matrix* mode of LindbladParameterizedGate
would construct an error generator in the std (matrix-unit) basis
and then change bases within the _construct_errgen function, which
is called repeatedly (b/c it's called by from_vector).  This is
unnecessary, as the basis transform can be performed once upon
construction and the error generator built in that basis (as was
already done in the sparse-matrix mode).  Now the dense-matrix-mode
does this as well.

Note that the generation function `std_error_generators` still
returns the generators in the *std* basis, to preserve similarity
with `lindblad_error_generators`, whose generators can be complex
even in normally-real (e.g. "pp") bases.