Adds "prep" and "effects" properties to GateSet.
These new properties give convenient access to the default (unique)
state preparation and POVM in a GateSet.  Unit tests are updated
to check that these work.

This commit also update the Intermediate Measurements tutorial to
demonstrate the ability to set POVM vectors directly, as well as
the new effects property.