Updates unit tests to improve coverage and fixes bugs. A WIP.
This is first commit in a series that improves the coverage of
the unit tests and fixes bugs as we go.  This commit updates many
files and constitutes the first pass through the source file up
to (and including) the "objects" sub-package.  Still lots of work
to do.  Coverage is about 85%.