Attempts to fix a couple of failing unit tests.
Changes minProbClip= argument from 1e-6 to 1e-4 in testCore to
try to prevent "finite Jacobian has infinite norm" assertion errors
when running MLGST.

Changes "link_to" argument from ('pkl','tex') to just ('pkl',) when
matplotlib is not installed since when 'tex' is given figures are
rendered as latex, which needs matplotlib.  In the future, we may
want to change this behavior so figures are only rendered in 'pdf'
mode.  See 'render_as_html' in merge_helpers.py for this rendering
logic.