Adds "manual transition" LinlogColormap factory for drift plots.
A new classmethod constructs a linear-logarithmic colormap where
the user can specify the point at which the gray/color transition
should occur, instead of specifying a number of boxes, dof, and
percentile.  Really the new factory method should be what is in
__init__ and the color-box-plot specific constructor a separate
factory, but we'll leave this as a future update.