Updates Sphinx-build conf.py to try to get readthedocs to work again.
At some unknown point in the past readthedocs builds started producing
nothing (but still 'passing' from a RTD standpoint).  This seems to
be caused by the RTD script's inability to 'import pygsti', even though
the setup.py install seems to have worked OK.  The best guess at the
problem currently is that conf.py needs to append the pygsti package's
path to sys.path.  This commit is a first stab at that.