Adds minor improvements to do_long_sequence_gst.
Uncomments a mistakenly-commented couple lines that automatically
set the 'comm' used for gauge optimization to the one passed to
do_long_sequence_gst (so gauge opt uses multiple procs!).

Secondly, this attempts the set the initial post-LGST gauge-opt
tolerance (whose default is the low 1e-8) based on the 'tol'
element of the gaugeOptParams arguement of do_long_sequence_gst.