Updates doc-generation conf.py to suppress lots of stdout.
RTD is hanging again after enabling the full documentation (builds
time out after 15min but only take 3min locally).  This update
uses code in conf.py to monkey-patch and hack Sphinx so that
warning messages (which we have a lot of) are suppressed.  Hopefully
this will allow the build to work.