Updates conf.py to try to disable warnings on RTD.
RTD builds are still hanging.  Next guess for what's up is that
since Sphinx 1.6.5 (what RTD currently uses) seems to use the root
"logger" of the python "logging" module we're not actually disabling
warning messages on RTD (this is corroborated by a prior failure
when numpydoc wasn't found).  This update to conf.py attempts to
remove the "warning handler" of the *root* logger if it doesn't
find one specifically for the "sphinx" name.