Finishes (at least for now) RTD debugging!
Here's the deal:  Read-the-docs builds were timing out.  With the
current conf.py's mix of monkey-patching and stdout suppression
the docs will build, but are *very* close to timing out at the 15min
limit.  Adding ":hidden:" tags to my_autosummary_class.rst (in a
prior commit) helped speed things up a little (this also fixed the
duplicated methods and attributes sections on a class page).  If
the builds timeout in the future, this commit adds a new file:
my_autosummary_class_inline.rst which generates all class member
documentation in a single "class page" rather than creating
individual stub-pages as my_autosummary_class.rst does.  This
approach is much faster and should bring build times down.

The current "patched_autosummary.py" has a slighlty more sophisticated
filter for "membership" than before (to reduce build time) - now
top-level pygsti.XXX routines are not documented (pygsti.XXX is NOT
seen as a member of pygsti when the actual object is, for instance
pygsti.drivers.longsequence.XXX) but 2nd-level routines are
(pygsti.drivers.XXX *is* a member of pygsti.drivers when the
actual object is pygsti.drivers.longsequence.XXX).  This logic is
found in the get_members function in patched_autosummary.py.

Finally, it should be noted that there are many (now suppressed)
warnings that, if we fixed them, might speed up the builds.

At this point, however, it looks like we'll be able to get RTD
to create a decently complete set of docs.