Begins integration of Tim's circuit code into pygsti.
Adds Label, Circuit, and StateSpaceLabels classes.  Label instances
generalize the string gate/spam labels that are elements of
GateStrings, and allow a set (tuple) of state-space labels to be
carried along with a string label.  Circuit is derived from GateString
and adds structure (e.g. layers) to a sequence of gates.  The
StateSpaceLabels class encapsulates the labeling of a Hilbert-space
decomposition into a direct-sum of the tensor products of sub-spaces.
This better collects state-label calculations from gatesetconstruction.py
and gate.py, and in particular allows integers to be used as state-
space labels (treated as qubits).  In its current state, Circuits can
be constructed and simulated using pyGSTi's normal density matrix
simulation.  For user convenience, GateSet's automatically create
 EmbeddedGateMaps (maybe calculator specific instances in future)
 when a gate is assigned with specific state-space labels that it
acts upon.