HOTFIX update to setup.py, attempting to fix "module not found" error.
After normal pypi install, users are getting "module not found" error
due to missing modules in setup.py (I think).  This commit adds missing
sub-packages to the "packages" line in setup.py.

This commit also adds Tim as an author and adds a list(.) around
the filter(.) used in setup.py (a repeat on an update currently
on the develop branch).