Updates gate calcs to include new "Unitary" variants.
New UnitaryGateMatrixCalc and UnitaryGateMapCalc objects implement
essentially the same calcultions as GateMatrixCalc and GateMapCalc
except for the way in which final probabilities are computed.  The
new classes assume gates are unitaries, so probabilities are
computed as |<E|G|rho>|^2 (where elements can be complex) instead of
 just <<E|G|rho>> as before (where elements are real).

Updates to the derivative computations in gatematrixcalc.py have
 not been worked out yet (NotImplementedErrors will be raised);
see comments in a TODO-marked section for how to proceed with this
in the future.