Adds enchancements to ProcessorSpec interface.
This commit polishes up some of the recently incorporated code,
making the interface a bit nicer and more logical.  Now the
ProcessorSpec object holds its clifford gateset as just another
model (ps.models['clifford']) instead of as as separate gate set
(ps.gateset).  The initialization of ProcessorSpec has been
restructured so the initial models and compilations to be constructed
are passed in as tuples of strings, and methods 'add_std_model' and
'add_std_compilation' can be used for adding more models &
compilations later on.

This commit also adds the 'get_clifford_symplectic_reps' function
to GateSet - where it should be - moving it out of it's temporary
location within the CompilationLibrary object.

Finally, the n-qubit GateSet-construction routines now take a
'on_construction_error' argument which can be 'raise', 'ignore',
or 'warn'.  This specifies the behavior when a gate cannot be
constructed, and allows one to construct a clifford-parameterized
gate set out of only the clifford gates within some larger set
(the non-clifford gates are just left out by setting 'warn' or
'ignore').