Updates Circuit to perform lazy evaluation of base members.
GateString now holds actual data in _tup and _str members
and access is mediated by @property functions.  This allows
these functions to be overriden in Circuit so that a Circuit
only needs to compute its underlying _tup when it's needed and
not after every change to the circuit (which occurs frequently
in compilations and so this alleviates a major bottleneck).