Updates Cython version of prs_as_polys.
Now fast_prs_as_polys replicates the SPAM-enabled and collapse()-free
calculations performed in the slower python version.  Comparison
between the python and cython versions show agreement (and with/without
the "fastmode" flag set to True).  So everthing looks good and there
are no known bugs in the prs_as_poly code as of now.  Running tests
shows that the term-based calc doesn't optimize logL as well for
longer lengths -- it's unclear whether this indicates a problem
somewhere or is due to the limited order of the probabilities.