Updates fast_prs_as_polys to use integers as gate labels.
Previously we just extracted the ".name" member of the gate Label
objects used to label the gates.  This was just a temporary solution,
as gates with the same name and different Labels wouldn't be
distinguished properly.  Dealing with strings in Cython is also a
pain.  This commit adds conversion (in Python) from Labels
to integers before calling the Cython routine, making the Cython
slightly faster and simpler.