Begins addition of clifford-simulation support.  Breaks calcs!
Incorporates "clifford" mode into GateMapCalc but in so doing alters
the convention for what a SPAMVec.toarray() returns.  Previously
it returned a (dim,1)-shaped array; now we want it to return just a
1D (dim,)-shaped array.  This is beneficial is several ways, but
breaks calculator functions until we get the calculator classes
fully updated.  But - I want to merge in recent changes in other
branches before really digging into that to avoid merge conflicts
down the line.