Fixes bugs and finishes initial Clifford calc (not working yet).
This commit fixes minor bug & gaps so that many of the calculator
variants work now.  The main addition, other than the Clifford
calculation work, was to implement 'deriv_wrt_params' for a
TensorProdSPAMVec (needed when using a matrix-based calc with
"reduced" n-qubit models).  The Clifford calcultion capability has
been completely plumbed up for simple circuit simulations, and
completes without error - but it need to be debugged as it doens't
give correct probabilities yet.  So this is all still a WIP.