Adds support for "H+S clifford terms" parameterization.
This new parameterization is implemented within the various
convert(...) routines within gate.py, spamvec.py and povm.py. Along
with this new parameterization, TensorProductSPAMVec is updated
to implement toarray() for prep stabilizer vecs and the "outcomes"
property for stabilizer effect vecs.  It also utilizes the new
symplectic_kronecker(...) function within get_order_terms(...) to
be able to construct tensor-product terms for stabilizer prep/effects
*without* needing to call RankOneTerm.collapse_vec(), which isn't
implemented for stabilizer vecs (and can't be for the effects).

A related update: convert(...) routines are now used for the
default_param processing in gateset member dictionaries.

Another related update:  GateTermCalc is updated to handle clifford
terms - by adding a flag for clifford/stabilizer term evolution
and by adding support for LindbladTermSPAMVecs with stabilizer
 effect vectors.  The effect-vector processing must be done slightly
differently because we can't "back-propagate" a StabilizerEffectVec
like we can a typical SPAMVec -- so we need to propagate the state
with the *adjoint* operators corresponding to the LindbladTermSPAMVec
terms.