Integrates StabilizerFrame propagation; clifford-term calcs work!
This commit updates clifford-term calculations to use StabilizerFrame
objects, and verifies, for the case when SPAM ops are not parameterized,
that these calculations work correctly.  That is, the produced
polynomials for computing probabilities agree in our test cases, as do
the logL/chi2 optimization steps and final result.  A few notes:

- To get the clifford-term optmization to work, we need to "kick" the
  starting point away from the target GateSet slightly.  This is because
  the clifford calcs are so exact that the symmetries present in the
  exact target gateset cause zeros in the (analytic) Jacobian that cause
  the optimizer to get stuck and never move away from the target.  This
  problem goes away when either kicking the starting point or falling back
  to finite difference Jacobian calculation, where the tiny amount of
  curvature picked up by the FD "derivative" is enough to get things started.

- spamvec.py and povm.py have hardcoded "DEBUG!!!" lines that remove
  SPAM vector parameterization (in order to compare between different
  calculation methods when we don't have a LindbladParameterizedSPAMVec
  object).  These lines need to be removed.  If errant behavior is seen
  in simulations later on, as suspect would be the SPAM-terms, and the
  adjoint_acton in particular, as this is only used for clifford-term
  calcs with imperfect effect vecs.

- all verification has been done with "fastmode = False" for computing
  term-calc polynomials.  That fastmode=True works the same as =False
  has been checked previously, so this shouldn't be an issue.

This commit marks what I expect is a "high water mark" in terms of
functionality.  Leading up to this point we've been adding term-based
and stabilizer-state calculation functionality and trying to verify
it's correctness.  Now this functionality exists, and following
work will likely focus on cleaning up the OLD & DEBUG portions,
refactoring, optimization, and improving programmer interfaces.