Begins addition of "_evotype"; Major gate.py refactor.
Old gate.py is still kept around as oldgate.py.  Current code is
updated so GateSetMember objects have a new member "_evotype",
which with specify the evolution type for the object.  Allowed values
are "densitymx", "statevec", "stabilizer", "svterm", and "cterm" as
of now.  The gate.py file has been updated: _evotype support was
added, but the file is refactored to consolidate duplicate code and
remove unnecessary functions from derived Gate classes.  Some
highlights are:
 - GateMap class removed (nothing different btwn it and Gate)
 - todense() and tosparse() functions allow lots of default logic
   to get pushed all the way up to Gate, making GateMatix and
   further downstream classes much leaner.  For example, almost all
   copy functions are gone; and functions like "depolarize", when
   not implemented, can just fall back to the base class versions
   nicely now.
 - ComposedGate derives from ComposedGateMap
 - EmbeddedGate derives from EmbeddedGateMap, and EmbeddedCliffordGate
   functionality has been folded into EmbeddedGateMap (when _evotype
   = "stabilizer").
 - BIG: LindbladParameterizedGate derives from
   LindbladParameterizedGateMap, and LindbladTermGate functionality
   is likewise folded in.  Some core functionality for constructing
   Lindblad gates has been pushed into gatetools.py to facilitate
   sharing with a future LindbladParameterizedSPAMVec.

Lots of things are broken at this point.  The standard densitymx
evolution was debugged and seems to work, but we'll need to add
evotype support to spamvec.py before we can really debug the other
modes.