Updates SPAM vecs, POVMs, and Instruments to use/track "evotype".
Use of the new "_evotype" member of the GateSetMember class has
been propagated to all the other (gates were done in prior commit)
type of GateSetMembers.  This commit also refactors spamvec.py a
bit, similar to the previous commit's refactoring of gate.py:
common method logic has been pushed to base class methods, reducing
the amount of redundant or unnecessary code in the derived classes.
One particularly noteworthy update is the replacement of the
LindbladTermSPAMVec, which just was usable for term calcs, with
LindbladParameterizedSPAMVec, which is more general (it holds
a LindbladParameterizedGateMap inside it).

Note that most copy(...) operations - for SPAM vecs, POVMs, and
instruments, have been removed, as we now rely on GateSetMember's
generic _copy(...) implementation.  Currently, the old version of
spamvec.py is held as "oldspamvec.py" - but povm.py and instrument.py
have just been updated (so to see the old versions - e.g. with their
copy methods - you'll need to pull the commit prior to this one).

pyGSTi is importable ok at this point, but lots of the calc code is
broken, as updating the calculators is the next thing to do.  So
don't expect anything to work at this point.