Adds statevec (SV) C++ rep classes and SV_prs_as_polys.
This commit wires up the necessary Cython/C representations objects
to run state-vector propagation and term-based calcs using statevec
evolution.  The latter is known to still now work completely, and
none of it has been rigorously tested.  Lots of debug statements
and comparisons throughout the calcs - not a good place to pull ;)