Adds replib.py, a Python mirror of representation classes and fns.
Mainly a collection of logic from elsewhere, replib.py provides the
same functionality as fastreplib.pyx but using pure python (slower)
routines.  This alternate implementation can be used 1) when Cython
is unavailable and 2) as a check.  Simple (mostly 1Q) tests check
out for all calculation modes, so replib.py seems to be working at
this point.

At this commit, the old non-representation versions of the map and
term calculations still exist and checks are performed that these
match the new replib versions.  (Testing was done by simply removing
fastreplib.so.)  Look for "DEPRECATED REPS" for parts that can
probably be removed when we're satisfied the rep calcs are correct.