Adds docstrings for a bunch of recently added functionality.
Also fixes a bug in Polynomial.__str__ so polys print correctly,
and fixes an anticipated (not seen yet) bug in
QubitGraph.are_connected_nodes when the graph is directed.

Docstrings still need to be added for all the replib functions
as well as ProcessorSpec (which also needs to be upgraded).