Updates unit tests and fixes bugs caught by tests.  All tests pass.
This commit continues the process of getting ready to merge the
feature-terms branch into develop.  If updates many files so that
all of the unit tests now pass.  It also adds two new "test" files,
demoCalcMethods{1,2}Q.py which run successfully but take a long time
(1Q ~ 10min and 2Q ~ hours) so aren't prefix with "test" so the won't
run automatically.  A number of the updates *to* the tests involved
adding explict conversions from strings to Labels, we'd like to be
unnecessary in the future once a special Label (names StrLabel?) class
is created.

A couple of notes:
- even though all the unit tests pass, there are a number of recently
  added features/functionality that aren't tested yet (dealing with
  circuits, processor specs, evotypes, and new calcs/simtypes).
- debugging structures & comparisions are still in place for testing
  the recently added calcs;  Once we're sufficiently convinced the
  new versions are working we need to remove the old version - see
  "DEPRECATED REPS".
- ProcessorSpec in particular still needs docstrings and updates to
  it works with the latest underlying evotype/simtype structure.