Splits Label object into two specialization classes, LabelTup and LabelStr
A Label object is a string "gate name" with an optional tuple of
state-space-labels (e.g. ("Q0","Q1")) identifying which part(s) of
the total Hilbert space an object acts upon.  Previously the state-
space-labels could be None and the object treated like a string.
This is fine, except that hashing via a __hash__ function, even
if it just returns "hash(tupleOrStr)" is *much* slower than setting
the class's __hash__ member to "tuple.__hash__" or "str.__hash__".
This is the motivating - and currently only - reason for making two
separate Label-derived classes, LabelTup and LabelStr.