Updates nqubit sequence selection.  Seems to work now.
`create_nqubit_sequences` has been updated to take advantage of
additional structure so it's much faster now.  Also, it now
populates a `cache` which can be used to hold a persistent database
of results on "cloud" subgraphs.  At this commit, there's still
lots of debug statements/logic leftover from testing different
variants, which will likely be removed soon.  So if the sequence
selection breaks coming back to this commit might be useful for
debugging.