Removes old "acton" functionality that is replaced by representations.
This commit removes the comparison checking between the "new"
representation-based calculation methods and the "old" methods
which used mixed Python/Cython within "acton" methods of the
Gate/SPAMVec classes.  It also removes the old implementations,
at least for the most part.  One piece that is still lacking:
in GateMapCalc the _compute_pr_cache only works for density-
matrix calculations (the "DM" case).  The replib modules need
to replicate this functionality for SV and SB cases.