Fixes a bug in create_nqubit_sequences; adds fallback functionality.
`get_fidpairs_needed_to_access_amped_polys` was failing to find a
sufficient set of fiducial pairs in some 2Q cases.  Initially, I
thought this was due to a problem with the algorithm's assumption
that the prep parts of the synthetic idle fiducials pairs are always
sufficient to be the preps for a non-idle power of the same germ.
So, functionality was added to search over different prep fiducials
on the core qubits if the idle-preps didn't work.  Later, it was
found that the actual problem was that numpy's matrix_rank can be
fooled into saying a single all-zero row has "rank 1" (even when
the singular values is O(1e-16)).  Then when a second (nonzero)
row is added to the Jacobian, the rank is *still* 1 and doesn't
increment, so no 2nd rows are ever accepted when building up a
Jacobian.  This commit leaves the prep-varying code in as a
fallback, even though it was found to be unnecessary for all the
test cases we've tried.