Adds initial support for auto-gator and "parallel-gate" labels.
Adds a new LabelTupTup class allowing a (n-qubit) gate label to
be a tuple of sub-Labels.  Also, and more significantly, the
gatestring compiling, eval trees, and calculators have been updated
so they no longer assume all the possible gate labels are the keys
of the GateSet or calculators "gates" dictionary.  A GateSet now
holds a (so far unused) _autogator member which in the future will
hold something (a function or class probably) that is called upon
-- by the new _getgate method of GateCalc -- to generate a (n-qubit)
gate for a label that *doesn't* exist in the calculator's self.gates
dictionary.  In this way, the auto-gator can specify how to create
a the noisy gates from doing several gates in parallel.  This may
be as simple as just composing them, or it may involve monkeying
around with the "serial gate" internals to product a "parallel gate".