Fixes bug in Circuit dirty-flag processing logic.
Circuit.__init__, in order to increase performance, didn't init
the base GateStrings _str member (instead it sets it to None) BUT
set the dirty flag == False.  This can lead to situations where a
'None' circuit.str occurs when it shouldn't.

This commit adds a second string-dirty-flag that keeps track of
whether the ._str member needs re-initializing.  In this way the
performance gains should remain while fixing the None problem.