Adds initial functionality for saving cached terms for std gatesets.
Cache files, holding the pre-computed polynomials for term-based
calculations will be stored in the new .../construction/caches/
directory.  Each cache dictionary is stored as two separate keys &
values files to speed up loading.  The new Label.tonative() function
is called to render the keys, which contain SPAM labels and GateStrings
as native python objects that are able to be un-pickled faster than
complex custom objects.  Moreover, because the various Label classes
inherit from basic python types and defer their equality testing and
hashing, there's no need to covert the keys to actual Label/GateString
objects as the hashing is identical.  Values (Polynomials) are stored
in their "compact" form by using numpy's savez_compressed method,
which seems to be very fast.

Currently, the new stdtarget.py module has a method for creating
cache files for standard gatesets which we need to run once to
generate the actual cache files (to be added to the repo later).
Also, the stdtarget._copy_target(...) need to be wrapped by new
stdXXX.copy_target(...) functions - this is the way the user will
get a copy of a GateSet that uses a precomputed cache.  Only
std1Q_XYI currently has a copy_target(...) implemented.