Adds 'comm' arg and plumbing to logl & chi2 functions using evaltree_cache.
A bug was found whereby an assertion fails because in the course of
do_long_sequence_gst a split evaltree (for a comm w/size NP, say)
gets cached and handed down to likelihoodfns.logl or .logl_terms
but *without* the corresponding comm.  The default no-comm condition
(w/NP=1) is then not compatible with a split evaltree, which requires
at least as many processors as there are subtrees.  Adding this
'comm' plumbing from the logl & chi2 functions up through
Estimate.misfit_sigma is accomplished by this commit.