Fixes a bug in MapEvalTree.split which causes some splits to fail.
The min/max default "cost rate" used in the tree splitting algorithm
where to small, resulting in some failures to split a tree into the
desired number of subtrees (and therefore an assertion failure).  This
commit fixes this issue, though may be a bit too conservative - so
further consideration of MapEvalTree.split could be given in the future
if it appears to be taking too long, etc.