Adds support for computing approximate Hessians (for faster errbars).
ConfidenceRegionFactory.compute_hessian(...) now has a new
``approximate` argument, which defaults to False but when set to
True computes an approximate Hessian.  This approximation neglects
all terms which would require the Hessian of the probabilities w.r.t.
the gate set parameters (the slow part), and only uses the Jacobian.

This commit adds logl_approximate_hessian to likelihoodfns.py and
"approximateHessian" as an argument to chi2(...) in chi2fns.py.

It appears to run without failing in the updated unit test, but hasn't
been tested beyond this.