Adds term-calc GST optimizations.
This commit adds algorithms for differentiating compact-form
polynomials.  This brings with it several advantages:
1) the "cache" member of GateTermCalc can (and now does) just
   hold *compact* polys now (speeding up loading and reducing
   required memory).
2) these algorithms are amenable to fast C implementation, and
   Cython routines have been created which speed up the evaluation
   and differentiation of compact polynomials.

There's still more room for optimization, and debugging code is
still present in many places.  Caching of derivatives was thought
to be a potential speedup but now the differentiation is so fast
it's not a bottleneck anymore.  To be continued; committing
now to switch to other more urgent updates.  Overall, these changes
have caused a roughly 20x speed increase in 1-st order 2QGST.