Adds a quick fix for loading text-format multiQ dataset files.
The parse_dataline routine could mistakenly identify the gatestring
column as a count (b/c "counts" are allowed to have the form
colon-delim-outcome-labels:count and colons are allowed in gatestrings
so the string "Gx:1" for instance was through to be a count).  A
quick (maybe temporary?) fix is to disallow 'G's in outcome labels,
which doesn't seem like too much of a constraint.  Now any column
value with a 'G' in it cannot be a count (enforced in parse_dataline).