Fixes bug that prevented parsing <gatename>:<qubitlabel> labels.
Forgotten "+" in regular expression caused failure to read any
gate label with a qubit-label longer than 1 charater.  It also
requires the label to be lowercase, which generally seems sensible
so it can be separated from other 'G'/'I'/etc entities in a gate
string.  The use of 'Q0',etc labels gave enough reason to allow
'Q' in qubit labels.  As of now, 'Q' is the *only* capital letter
 allowed in the qubit label.