Adds lazy effect creation to TensorProdPOVM and StabilizerZPOVM.
This was motivated by wanting ProcessorSpec objects to initialize
faster, even when there are many qubits.  The bottleneck was the
creation of *all* the 2^nQubits effect vectors within a tensor-prod
or stabilizer-Z POVM.  Now these POVMs create effect vectors only
on demand (or when asked to `compile_effects` by a GateSet).  This
functionality has only been lightly tested.