Updates CompilationLibrary: allowed_filter and get_ methods.
Two get_* methods have been added to create but *not* store local
and nonlocal compilations.  This is important for processor specs
with a large number of qubits where you don't want to store all
possible non-local compilations (in particular).

A new 'allowed_filter' argument has been added to the functions for
adding (or just getting) nonlocal compilations, so that the user
can specify that only a subset of the gates containedin the library
are available for building nonlocal compilations from.