Adds "on-demand" computation of outcome probabilities.
Adds 'dataset' arguments to GateSet.compile_gatestrings and
upstream functions (like evaltree generation functions) so that
a DataSet can be specified at the compile step which restricts the
to-be-computed probabilities to only those for which there are
nonzero counts (i.e. observed data).  This is especially advantageous
in the many-qubit case when there are so many possible outcomes
that we can't possibly see them all (all data sets will be "sparse").