Adds better & support for "sparse" DataSets and makes them the default.
DataSet has been updated so that (at least most of) its add_* methods
have a "recordZeroCnts" argument that is False by default in
add_count_dict.  This means that any explicit 0-counts in added data
are ignored except for potentially adding a valid outcome label.
Resulting DataSets in many ways still look like they contain all
the possible outcome labels (though what is "possible" may be
defined just by what has been observed for any GateString), as
the DataSetRow.allcounts property gives a dictionary containg all
of dataset.get_outcome_labels(), and indexing a DataSetRow with
an outcome label that isn't present will just return 0.

This update closely interplays with the last commit of "on-demand"
gatestring compilation, as probabilities are then computed for all
outcomes *present* in the DataSet (even if they're zero).  Hence,
to realize the benefit of the "on-demand" compiling you need to
use "sparse" DataSets - ones whose rows only contain outcomes for
nonzero counts.

Note: There are potentially more fragile pieces in chi2fns.py and
likelihoodfns.py which index the DataSetRow.counts property
(a dict of *only* the nonzero outcome labels) with outcome labels
which may not be restricted. Given the last update to "on-demand"
gate string compiling, this shouldn't happen when the Dataset
used was also passed to GateSet.compile_gatestrings.  But when
compiled gatestrings are given to a function like logl_terms via
an eval tree this need not always be the case - hence the fragility.