Makes cntVecMx population more robust.
This commit addresses the fragility issue described in a recent
(maybe last) commit.  It replaces cnts[x] with cnts.get(x,0) in
a number of places, where cnts is a DataSetRow.counts dictionary.
The premise is that this counts dict may not always contain all
of the "theoretically possible" outcomes - just the ones observed,
and so we default to 0 (counts) when an outcome label is missing.
That said, this usually shouldn't be necessary, as the construction
of cntVecMx should only be using nonzero counts (if compile_gatestrings
was given an appropriate DataSet object).

One further note: not in this commit, but in the *last* one the
MultiDataSet.add_dataset method was revamped considerably to
deal with adding DataSets that don't have the exact structure
existing in the MultiDataSet.  This hasn't been tested yet though,
and should be mentioned somewhere in the commit log... so better
late than never.