Updates find_amped_polys_for_syntheticidle to fix nQ sequence construction.
The recent update to use a greedy algorithm in fidpair selection needs
to have the correct "wrtParams" variable sent to find_amped..., or
it will run into an assertion error that it can't amplify all the parameters.
This commit fixes this issue within the n-qubit sequence construction
routines (so they work again), and also adds a new "require_all_amped"
argument to find_amped_polys_for_syntheticidle (currently always True)
in case we don't want to require that all parameters be amplified in
the future.