Updates Circuit object to allow for labels with sslbls == None.
The Circuit object now treats a None state-space-labels part of a
gate label as meaning that the gate acts on all the qubits in the
circuit, but *doesn't modify* the gate labels - you have to do this
separately if you want to change the labels.  Previously it changed
the labels, adding the all-qubit sslbls, but this seems like an
unexpected (bad) side effect.  We might want to add a warning when
it does this in certain cases (like maybe when all the gate labels
have None sslbls?) in the future.