Updates _fas to use new Cython routines for "advanced" case.
The "fancy assignment" functionality implemented by _fas was
causing a bottleneck since switching over to sparse datasets.
This is because not-computing 0-count probabilities causes the
indices corresponding to a given spam-tuple (~outcome) to no
longer be neat slices (w/step = #outcomes typically).  This required
_fas to use numpy advanced indexing with additional copying and
index manipulation which is slow.  This commit adds fast_fas_helper_Xd
(for X=1,2,3) functions in fastcalc.pyx which do this advanced
assigment in an efficient way.

Observed speedup for 2Q term-based GST was ~10x, which just brings
it back to the speed it was running at before we added the on-demand
probability computation.