Updates fast_fas_helper_?d routines to work with non-contiguous arrays.
Now 2d and 3d routines (1d should always be contiguous) are able to
accept non-contiguous input arrays and use the .strides property of
numpy arrays to perform the correct assignments.

It was found in unit tests, that in particular for Hessian
calculations the `rhs` input array is not always contiguous.  In
addition to being faster than requiring a np.ascontiguousarray call,
this allows _fas to continue to modify arrays *in-place* and avoid
copying.