Fixes bug in _fas picked up by unit tests.  Updates unit test.
We had been using the new Cython fas helper functions, which only
assign (not add), even when add=True.  Now we check that add==False
before using the fast Cython functions - since we don't use
add=True much or even at all it doesn't seem worth implementing this
case in Cython (even though it would be trivial).

Use of the cython helpers in _fas also need to respond correctly
when single integers are used as indices.  For example, you can
assign a 2D matrix to a 3D destination when one of the indices is
just an integer.  _fas has been updated to properly account for
this, so that the Cython routines are always fed source and
destination arrays of the same dimension.