Fixes Lindblad-gate deriv_wrt_params bug (affects optimizations!)
An overzealous removal of numpy array copying was responsible for
this bug in _dexpSeries(...) and _d2expSeries(...) which ultimately
caused a LindbladParameterizedGate (the Map doesn't use this) to
be itself be modified when calling deriv_wrt_params or
hessian_wrt_params.  This ill effect was especially problematic
within LM optimizations, which evaluate the Jacobian and expect
the objective function to remain unchanged (!), ultimately resulting
in occasional failures of the optimizer.  This issue is fixed by
this commit.