Updates pyGSTi to work with plotly version 3.
In version 3 plotly changes some of its internal structure, which
requires an updated approach to making Figure objects pickle-able
(they still aren't without "help"!), and updates to use the new
hierarchy of objects (e.g. using graph_objs.layout.XAxis instead
of just graphy_objs.XAxis as done in v2).  The updated code checks
the plotly version for pickling and when using graph objects so
pyGSTi should be compatible with plotly v2.x.x and v3.x.x now.

Note that in the prior commit the custom "polarfixed" versions
of the plotly Javascript library were removed from the repo as now
we just include the latest version of plotly.js identical to the
one on the CDN.  This seems fine since we don't use polar plots
anymore, but could be an issue if the polar plots still aren't fixed
by the time we want/need them again.