Adds commented-out "import faulthandler"; a vestige from prior debugging.
The faulthandler package can be useful for debugging as it prints
a traceback when a segfault occurs.  The downside it that it only
prints a trace through python stack frames (and usually the segfault
occurs in Cython routines).  In any case, this commit adds the
code that was used in debugging, commented out at the top of
gatetermcalc.py, for potential future use.